EngineController: make trackPositionMs() return qint64 instead of int
Phonon returns position in qint64, every use of trackPositionMs()
computes with qint64, so there is no need to introduce artificial limit
of 24 days (stream) length.

+++ src/EngineController.h
@@ -83,7 +83,7 @@ public:
     int trackPosition() const;
 
     /** @return track position (elapsed time) in milliseconds */
-    int trackPositionMs() const;
+    qint64 trackPositionMs() const;
 
     /**
      * Returns the current track that is loaded into the engine.
